backend/requirements.txt
Django>=5.0.0,<5.2.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
drf-yasg>=1.21.7,<1.22.0
django-cors-headers>=4.3.1,<4.4.0
django-filter>=23.5,<24.0
Pillow>=10.1.0,<10.2.0
python-dotenv>=1.0.0,<1.1.0 